精简core模块，修复重置id后id不正确问题，优化单元测试。
修复：
1.修复重置id后id不正确问题。

改动：
1.core模块只保留核心算法和dao接口，迁移其他内容到server。
2.迁移dao实现的sql到xml。
3.单元测试执行时判断id是否重复。

+++ leaf-core/src/main/java/com/sankuai/inf/leaf/segment/model/LeafAlloc.java
@@ -27,7 +27,7 @@ public class LeafAlloc {
     /**
      * 当前业务允许的最大id，可以用来控制id长度。
      */
-    private int maxNumber;
+    private long maxNumber;
     /**
      * 描述。
      */
@@ -69,11 +69,11 @@ public class LeafAlloc {
         this.step = step;
     }
 
-    public int getMaxNumber() {
+    public long getMaxNumber() {
         return maxNumber;
     }
 
-    public void setMaxNumber(int maxNumber) {
+    public void setMaxNumber(long maxNumber) {
         this.maxNumber = maxNumber;
     }
